fix(examples): fix sveltekit-openai type errors
