+++ examples/sveltekit-openai/src/lib/components/ui/button/index.ts
@@ -1,17 +1,17 @@
-import Root, {
+import Root from './button.svelte';
+import {
+  buttonVariants,
   type ButtonProps,
   type ButtonSize,
   type ButtonVariant,
-  buttonVariants,
-} from './button.svelte';
+} from './variants.js';
 
 export {
   Root,
-  type ButtonProps as Props,
-  //
   Root as Button,
   buttonVariants,
   type ButtonProps,
+  type ButtonProps as Props,
   type ButtonSize,
   type ButtonVariant,
 };
